Split future_factory into a separate header
