+++ hpx/lcos/wait_any.hpp
@@ -139,6 +139,7 @@ namespace hpx
 
 #include <hpx/config.hpp>
 #include <hpx/lcos/future.hpp>
+#include <hpx/lcos/local/futures_factory.hpp>
 #include <hpx/lcos/wait_some.hpp>
 #include <hpx/runtime/threads/thread.hpp>
 #include <hpx/util/always_void.hpp>
